main_page & hitsory merge
